A big roll-up of changes over the past couple of years
I've not been maintaining very good git hygiene and have a stack of
changes that I've made locally, in part as an investigation and prototype
into maintaining compatibility with more recent versions of Android.

This is now getting to a degree of stability, so I think it's better to
have this in and start producing artefacts than waiting for it to be
finessed.

Things that have changed in this commit:
- A much improved UI that better guides the user through the file source
and target selection.
- Use of scoped storage to get access to the source and target. This has
been the most significant change and has been required to support later
versions of Android. It's had a performance impact because the API is
notoriously slow. There may be some opportunities to shortcut this.
- Extracting of the GOG game ID as part of validation and grabbing the
icon, background image and title to be displayed.
- A help dialog to guide the users to a help page if their Inno file cannot
be loaded. The page is incomplete.