Allow extracting directly to the filesystem
We might not want to use the document storage as an extract target. This
adds a second Extract method to the service that allows passing in a file
on the filesystem. It is assumed that this is accessible to the process. If
this method is used, instead of a temporary file being used, the extraction
is done directly to the filesystem.